🚚 Move components to Widgets
